+++ tf_encrypted/layers/tanh_test.py
