+++ tf_encrypted/protocol/pond/truncate_test.py
