+++ tf_encrypted/protocol/securenn/argmax_test.py
