+++ tf_encrypted/protocol/strided_slice_test.py
